incident_ predictor_backend/requirements.txt
asgiref==3.2.3
cycler==0.10.0
Django==3.0.2
djangorestframework==3.11.0
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
pandas==0.25.3
prettytable==0.7.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.20.1
scipy==1.4.1
seaborn==0.9.0
six==1.14.0
sqlparse==0.3.0
